requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery[redis]==5.3.4
kombu==5.3.4
redis>=4.5.2,<5.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Reddit API
praw==7.7.1

# AI/ML dependencies
openai==1.3.7
numpy>=1.24.0,<2.0.0

# Ghost CMS
PyJWT==2.8.0
requests==2.31.0
httpx==0.25.2
tenacity==8.2.3

# Security (Vault removed for MVP)
cryptography==43.0.3

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Image processing
Pillow==10.1.0

# Template engine
pybars3==0.9.7
jinja2==3.1.2
python-multipart==0.0.6
markdown==3.5.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
psutil==5.9.6